events-service/requirements.txt
grpcio>=1.65.0
grpcio-tools>=1.65.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
kafka-python>=2.0.2
protobuf>=4.25.0